+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Change `RadioButton` and `CheckBox` label click.
+- Feature: Add `direction` on all `ArrowIcon` component.
+- Fix: IconSearch prop in Search component.
 - Feature: Add default icon sizes.
 
 ## [14.0.3] - 2017-09-13